services/check_respondent_identity/requirements.txt
Flask==3.1.0
google-cloud-bigquery==3.27.0
google-cloud-bigquery-storage==2.27.0
gunicorn==23.0.0
twilio==9.4.1
Werkzeug==3.1.3
